Handle all requirements for including attestations in blocks.
